Refactored grok pattern for parsing proxy data
The original contribution was specific to the contributor situation (apparently amavisd with postfix on re-injection). As we should not depend on being postfix being the after-proxy re-injection system, or amavisd being the proxy itself, I changed it to be more generic.
The new pattern does try to extract smtp status codes that the proxy generated from the log line, but leaves the proxy result message otherwise untouched. Implementors should add a simple pattern downstream that parses their specific output as captured in `postfix_proxy_status_message`, when necessary.